+++ app/helpers/tasks_helper.rb
@@ -1,2 +1,3 @@
 module TasksHelper
+  include Pagy::Frontend
 end