+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # HEAD
 
+- feat: Tap h1 to reload
+
 # v0.4.0 (2018-11-13)
 
 - feat: Show 'connecting...' when network is slow